fix: 컨트롤러가 요청의 sortBy 값을 SortBy enum으로 읽어들이지 못하는 문제 해결
SortBy enum의 값을 상수의 표현 형태인 대문자로 표기하다보니, 실제 요청에서 들어오는 값인 소문자 값과 매칭이 되지 않는 문제가 있었음. 예를 들어, sortBy=rating 형태로 들어온 요청 파라미터를 컨트롤러가 SortBy.RATING 으로 변환하지 못하고 있었음.

SortByConverter를 정의하여 컨트롤러가 정상적으로 파라미터 변환을 할 수 있도록 함. 해당 컨버터를 등록할 WebMvcConfig는 모든 상황에서 동작해야하므로, QueryLoggingInterceptor를 사용하는 기존 WebMvcConfig는 DevWebMvcConfig로 변경함.
