+++ src/main/java/com/jvnlee/catchdining/common/config/DevWebMvcConfig.java
@@ -1,0 +1,22 @@
+package com.jvnlee.catchdining.common.config;
+
+import com.jvnlee.catchdining.common.interceptor.QueryLoggingInterceptor;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Profile("dev")
+@Configuration
+@RequiredArgsConstructor
+public class DevWebMvcConfig implements WebMvcConfigurer {
+
+    private final QueryLoggingInterceptor queryLoggingInterceptor;
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(queryLoggingInterceptor);
+    }
+
+}
