+++ src/main/java/com/jvnlee/catchdining/common/config/WebMvcConfig.java
@@ -1,21 +1,20 @@
 package com.jvnlee.catchdining.common.config;
 
-import com.jvnlee.catchdining.common.interceptor.QueryLoggingInterceptor;
+import com.jvnlee.catchdining.common.web.SortByConverter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.format.FormatterRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-@Profile("dev")
 @Configuration
 @RequiredArgsConstructor
 public class WebMvcConfig implements WebMvcConfigurer {
 
-    private final QueryLoggingInterceptor queryLoggingInterceptor;
+    private final SortByConverter sortByConverter;
 
     @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(queryLoggingInterceptor);
+    public void addFormatters(FormatterRegistry registry) {
+        registry.addConverter(sortByConverter);
     }
+
 }
