+++ src/main/java/com/jvnlee/catchdining/domain/restaurant/model/SortBy.java
@@ -1,5 +1,5 @@
 package com.jvnlee.catchdining.domain.restaurant.model;
 
 public enum SortBy {
-    RATING, REVIEW_COUNT
+    RATING, REVIEWCOUNT
 }
